Remove wait_postmaster_connection
Rely on default pg_ctl behavior that is to wait for the operation to
complete. Maybe the user has to tweak the maximum number of seconds to
wait for the pg_ctl modes (start, stop, ...). The default is 60 seconds
that can be too small for some catchup scenarios. The user can set the
PGCTLTIMEOUT environment variable to avoid an undesirable behavior. (I'm
not sure if we need a --timeout option or document the PGCTLTIMEOUT
here.)