pg_ctl: add a value to timeout option that waits forever
By default, pg_ctl waits for a certain amount of time (--timeout) when
waiting for an operation to complete. If a negative value is specified,
pg_ctl will wait forever (until the operation is completed or it is
canceled by the user).

There are certain scenarios that a timeout is undesirable. In particular
when you are using pg_ctl in a script/program that should wait until the
operation is complete (start and promote actions can take varying
amounts of time depending on the write-ahead log it needs to process,
hence, even using a big number for timeout seems unreliable and
ungraceful).