log_min_messages per backend type
Change log_min_messages from a single element to a comma-separated list
of elements. Each element is BACKENDTYPE:LOGLEVEL. The BACKENDTYPE are
archiver, autovacuum (includes launcher and workers), backend, bgworker,
bgwriter, checkpointer, logger, slotsyncworker, walreceiver, walsender,
walsummarizer and walwriter.

The old syntax (a single log level) is still accepted for backward
compatibility. In this case, it means the informed log level is applied
for the backend types that are not informed. The default log level is
the same (WARNING) for all backend types.