Fix design flaw
The publications should be created *before* the consistent LSN. The
original idea was to create the publications after the promotion but
there might be transactions between the consistent LSN and the
transaction that create the publication object. In this case, when it
starts decoding the first transaction, it doesn't see the publication
yet (that's because it was created after it.)

Fix it by creating the publication objects while creating the
replication slots.

https://www.postgresql.org/message-id/TY3PR01MB9889C5D55206DDD978627D07F5752%40TY3PR01MB9889.jpnprd01.prod.outlook.com