Refactor primary_slot_name code
We used to have a use_primary_slot_name() function to verify that a
replication slot is in use for physical replication. However, with the
new setup_publisher / setup_subscriber functions, we split the referred
function into 2 pieces to avoid extra connections. It also simplifies
the code a bit.

It also fixes an issue in setup_subscriber that was referring to the
wrong row / attributes.