Get absolute path from pg_ctl and pg_resetwal
It basically reverts commit b65790ac630ec616df9aefe84d04f80bcc69ed53
that unified it and used a base directory. The reason is to provide
specific error message for each of the binaries. A generic function
(get_exec_path) will check if the binary is available in the same
directory as pg_createsubscriber and it has the same version. It returns
an absolute path for the binary that is used by some functions.