bug: remove publication on subscriber
made_subscription is not the right condition to drop publications on
subscriber. When the recovery ends, publications should have be
available on subscriber. However, if this program ends before creating
subscriptions, it doesn't have a chance to set made_subscription. Hence,
the publications on subscriber won't be removed accordingly.

This patch uses a new variable (recovery_ended) to control if the
publications on subscriber should be removed. It closes the gap
explained in the previous paragraph.