Replace subscriber-server option
This is a slightly modified version of v22-0004. It basically removes
the subscriber-server option and add 3 new options: subscriber-port,
subscriber-username and socket-directory. The main reason is to prevent
external access during the execution (that should minimize the chance of
failure). The new options allow you to connect to the target server and
execute the required queries, although, it doesn't allow you to specify
the majority of the recognized parameters. That's ok for the limited
time this tool will run.