refactor: function signatures
Instead of returning a boolean to validate if the function runs
successfully, include the error test inside the function. It helps while
inspecting the code because you don't to have to check outside the
function body.

It also fixes some places that forgot to close the connection before
exit(1). The number of connection attempts during recovery was fixed to
finish after NUM_CONN_ATTEMPTS.