Extra stop and start server calls
It fixes commit 57dcbcf788 that replaced subscriber-server option with
socket directory, port and username options. The server must start with
the informed options *before* start the validation steps. Tests didn't
catch it because it was using the same connection parameters to start
the server and to pass to the command-line options.