Add new options to assign object names
The publication, subscription and replication-slot options allow the
user to assign names to the referred objects. Use multiple options if
you have more than one database. If you specify 3 databases and decide
to assign publication names, you must specify 3 publication names. The
same is valid for the other options too. If the option is not provided,
it assigns a generated name to the object.