Separate GUC for replication origins
The new GUC (max_active_replication_origins) defines the maximum number
of active replication origins. The max_replication_slots was used for
this purpose but it is confusing (when you are learning about logical
replication) and introduces a limitation (you cannot have a small number
of replication slots and a high number of subscriptions).

For backward compatibility, the default is -1, indicating that the value
of max_replication_slots is used instead.